write tes for reset password and rename routes and just edit rename route to check rename to does not have a duplicated.

+++ server/util/rename.ts
@@ -6,7 +6,9 @@ const __cwd = process.cwd();
 const rename = (cwd: string, fileName: string, newName: string) => {
   try {
     const basePath = path.join(__cwd, "/server/static/cloud/", `${cwd}/`);
-
+    if (fs.existsSync(basePath + newName)) {
+      return { status: false, msg: "the name is exist" };
+    }
     fs.renameSync(basePath + fileName, basePath + newName);
 
     return { status: true, msg: "renamed correctly" };
